requirements.txt
click==8.1.3
distlib==0.3.5
eli5==0.13.0
fastjsonschema==2.16.1
filelock==3.7.1
Flask==2.1.0
graphviz==0.20.1
itsdangerous==2.1.2
Jinja2==3.1.2
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
lightgbm==3.3.2
MarkupSafe==2.1.1
nbclient==0.6.6
nbconvert==6.5.0
tensorflow-cpu
nbformat==5.4.0
nest-asyncio==1.5.5
pandas-datareader==0.10.0
platformdirs==2.5.2
pmdarima==1.8.5
requests==2.28.1
SCALECAST==0.13.3
sklearn==0.0
tabulate==0.8.10
tinycss2==1.1.1
traitlets==5.3.0
virtualenv==20.16.2
Werkzeug==2.2.1
wincertstore==0.2
yfinance==0.1.74
